+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+In next release ...
+
+- Fix an issue where `auto_reload` is enabled and a file is loaded
+  from a package that wasn't zip-compressed.
+  (`#402 <https://github.com/malthe/chameleon/issues/402>`_)
+
+
 4.4.2 (2023-12-18)
 ------------------